# app/openapi.yaml
openapi: "3.1.0"

info:
  title: SpoolBeacon API
  version: "1.0.0"
  description: >
    REST API for SpoolBeacon filament inventory.
    Authenticate via **API Keys** (`/api-keys` page, admin only).

servers:
  - url: /api/v1
    description: Current instance

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string

    Manufacturer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]

    ProductSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        manufacturer:
          type: string
        manufacturer_id:
          type: integer
        material:
          type: string
        color_name:
          type: string
        color_hex:
          type: ["string", "null"]
        color_name_2:
          type: ["string", "null"]
        color_hex_2:
          type: ["string", "null"]
        diameter_mm:
          type: number
        nominal_weight_g:
          type: integer

    Spool:
      type: object
      properties:
        id:
          type: integer
        spool_code:
          type: string
        product_id:
          type: integer
        status:
          type: string
          enum: [new, opened, almost_empty, empty, archived]
        initial_weight_g:
          type: number
        remaining_weight_g:
          type: number
        fill_percent:
          type: number
        storage_location:
          type: ["string", "null"]
        storage_status:
          type: string
        opened_at:
          type: ["string", "null"]
          format: date-time
        last_dried_at:
          type: ["string", "null"]
          format: date-time
        created_at:
          type: ["string", "null"]
          format: date-time
        product_name:
          type: string
          description: Only present when loaded via /spools endpoints
        manufacturer:
          type: ["string", "null"]
          description: Only present when loaded via /spools endpoints
        material:
          type: string
          description: Only present when loaded via /spools endpoints
        color_name:
          type: string
          description: Only present when loaded via /spools endpoints

    PrintJobLine:
      type: object
      properties:
        id:
          type: integer
        spool_id:
          type: ["integer", "null"]
        spool_code:
          type: string
        product_name:
          type: string
        used_g:
          type: number

    PrintJob:
      type: object
      properties:
        id:
          type: integer
        job_code:
          type: string
        print_name:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        printed_at:
          type: string
          format: date-time
        total_used_g:
          type: number
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PrintJobLine"

    PrintJobCreate:
      type: object
      required: [lines]
      properties:
        print_name:
          type: string
          description: Optional name for the print
        notes:
          type: string
        lines:
          type: array
          minItems: 1
          items:
            type: object
            required: [spool_id, used_g]
            properties:
              spool_id:
                type: integer
              used_g:
                type: number
                exclusiveMinimum: 0

    SpoolPatch:
      type: object
      required: [remaining_g]
      properties:
        remaining_g:
          type: number
          minimum: 0
          description: New remaining weight in grams

    PaginatedPrints:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/PrintJob"

paths:
  /health:
    get:
      operationId: health
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /manufacturers:
    get:
      operationId: listManufacturers
      summary: List all manufacturers
      responses:
        "200":
          description: Array of manufacturers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manufacturer"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /products:
    get:
      operationId: listProducts
      summary: List all filament products
      responses:
        "200":
          description: Array of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /products/{id}:
    get:
      operationId: getProduct
      summary: Get product detail with spools
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product with spools array
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ProductSummary"
                  - type: object
                    properties:
                      spools:
                        type: array
                        items:
                          $ref: "#/components/schemas/Spool"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /spools:
    get:
      operationId: listSpools
      summary: List spools
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, opened, almost_empty, empty, archived]
          description: Filter by spool status
      responses:
        "200":
          description: Array of spools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spool"
        "400":
          description: Invalid status filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /spools/{id}:
    get:
      operationId: getSpool
      summary: Get a single spool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spool object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spool"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchSpool
      summary: Update spool remaining weight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpoolPatch"
      responses:
        "200":
          description: Updated spool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spool"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /prints:
    get:
      operationId: listPrints
      summary: List print jobs (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated print jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPrints"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPrint
      summary: Log a new print job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintJobCreate"
            example:
              print_name: Benchy
              lines:
                - spool_id: 1
                  used_g: 6.2
      responses:
        "201":
          description: Print job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error (e.g. used_g exceeds remaining weight)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"